==> notebook/01_preprocessing.ipynb <==
# 01 Preprocessing

import glob
import pandas as pd
import scanpy as sc
from natsort import natsorted

## _in vivo_ & _in vitro_

metadata = pd.read_excel("../data/raw/metadata.xlsx", index_col=0)
metadata.index = metadata['Well_ID']
metadata.SubGroup = metadata.SubGroup.astype(str)

files = [f"../data/raw/counts/{batch}.txt" for batch in natsorted(metadata.Amp_batch_ID.unique())]
counts = pd.concat([pd.read_table(x, sep=" ") for x in files], axis=1).T

metadata = metadata.loc[counts.index]

adata = sc.AnnData(counts, obs=metadata)
adata.write("../data/processed/01_raw_counts.h5ad")

## SB12

metadata = pd.read_excel("../data/raw/metadata_SB12.xlsx", index_col=0)
metadata.index = metadata['Well_ID']

files = [f"../data/raw/counts/{batch}.txt" for batch in natsorted(metadata.Amp_batch_ID.unique())]
counts = pd.concat([pd.read_table(x, sep=" ") for x in files], axis=1).T

metadata = metadata.loc[counts.index]

adata = sc.AnnData(counts, obs=metadata)
adata.write("../data/processed/06_rawcounts.h5ad")

!jupyter nbconvert --to html 01_preprocessing.ipynb --output-dir="../reports"
